feat: add standard error codes and mapping for consistent error handling
