+++ packages/domain/src/index.ts
@@ -2,6 +2,7 @@ export * from "./category.js";
 export * from "./user.js";
 export * from "./config.js";
 export * from "./product.js";
+export * from "./error-codes.js";
 export * from "./common.js";
 export * from "./shared.js";
